+++ packages/components/.storybook/main.ts
@@ -1,19 +1,16 @@
-import type { StorybookConfig } from "@storybook/react-vite";
+import type { StorybookConfig } from "storybook-react-rsbuild";
 
 const config: StorybookConfig = {
   stories: ["../stories/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
   addons: [
+    "@storybook/addon-links",
     "@storybook/addon-essentials",
-    "@storybook/addon-onboarding",
-    "@chromatic-com/storybook",
-    "@storybook/experimental-addon-test",
+    "@storybook/addon-interactions",
+    "storybook-addon-rslib",
   ],
-  core: {
-    builder: "@storybook/builder-vite",
-  },
   framework: {
-    name: "@storybook/react-vite",
+    name: "storybook-react-rsbuild",
     options: {},
-  },
+  }
 };
 export default config;