+++ packages/components/stories/Button.stories.tsx
@@ -1,7 +1,8 @@
-import * as React from "react";
+import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import { Button } from "./../src/components/ui/button";
 import "./../src/index.css";
+import { Button } from "../dist";
+
 
 const meta: Meta<typeof Button> = {
   title: "Design System/Button",
@@ -11,19 +12,21 @@ const meta: Meta<typeof Button> = {
 type Story = StoryObj<typeof Button>;
 
 export const Primary: Story = {
-  render: () => <Button className="text-red-500 bg-blue">Button</Button>,
+  render: () => <Button size='default' variant='default' className="text-red-500 bg-blue">Button</Button>,
 };
 export const DestructiveButton: Story = {
-  args: {
-    variant: "destructive",
-    content: "Destructive",
-  },
+  render: () => (
+    <Button size='default' variant="destructive" className="text-red-500 bg-blue">
+      Destructive
+    </Button>
+  ),  
 };
 export const DisabledButton: Story = {
-  args: {
-    disabled: true,
-    content: "Disabled",
-  },
+  render: () => (
+    <Button size='default' variant='default' disabled className="text-red-500 bg-blue">
+      Disabled
+    </Button>
+  ),
 };
 
 export default meta;