+++ packages/components/changelog.md
@@ -5,6 +5,11 @@ All notable changes to this component library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.0.35 - (2025-04-08)
+
+- Migrate from vite to rslib.
+- Update calendar component from v8 to v9.
+
 ## 0.0.34 - (2025-03-27)
 
 - Upped package versions.